demo-ui/requirements.txt
flask
requests
google-cloud-storage
